fixed bug in settings, fixed bug in accountlistpage, reworked update function

+++ lib/screens/splash/cubit/splash_screen_cubit.dart
@@ -1,5 +1,9 @@
 import 'package:bloc/bloc.dart';
+import 'package:flutter/foundation.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
+import '../../../repository/main_repository.dart';
+import '../../../repository/preferences_repository.dart';
+import '../../../repository/settings_repository.dart';
 import '../../../services/update_service.dart';
 
 part 'splash_screen_state.dart';
@@ -7,18 +11,43 @@ part 'splash_screen_state.dart';
 part 'splash_screen_cubit.freezed.dart';
 
 class SplashScreenCubit extends Cubit<SplashScreenState> {
-  SplashScreenCubit() : super(const SplashScreenState.initial());
+  final MainRepository mainRepository;
+  final SettingsRepository settingsRepository;
+  final PreferencesRepository preferencesRepository;
+
+  SplashScreenCubit({
+    required this.mainRepository,
+    required this.settingsRepository,
+    required this.preferencesRepository,
+  }) : super(const SplashScreenState.initial());
 
   void initialize() async {
+    emit(SplashScreenState.checkingForUpdates());
     try {
-      await UpdateService().deleteUpdateFiles();
-      UpdateService().checkForUpdate().then((value) async {
-        if (value) {
+      if (!kDebugMode) {
+        await UpdateService().deleteUpdateFiles();
+        final updateNeeded = await UpdateService().checkForUpdate();
+        if (updateNeeded) {
           emit(const SplashScreenState.updateAvailable());
-        } else {
-          emit(const SplashScreenState.initialized());
+          await Future.delayed(const Duration(seconds: 3));
+          UpdateService().updateApp();
+          return;
         }
-      });
+      }
+
+      final wowPath = await preferencesRepository.getWowPath();
+      final dataPath = await preferencesRepository.getDataDirectoryPath();
+      settingsRepository.secondsToWaitForGameToStart =
+          await preferencesRepository.getWaitTillGameStarts() ?? 3;
+      if (wowPath != null) {
+        settingsRepository.fillWithExecutablePath(wowPath);
+
+        settingsRepository.wowDataDirectory = dataPath;
+        settingsRepository.wowRealmFilePath = '$dataPath/realmlist.wtf';
+      }
+      emit(const SplashScreenState.initialized());
+
+      ;
     } catch (e) {
       emit(SplashScreenState.failed(errorMsg: e.toString()));
     }
